changed signs in mzpinstru_to_mzpsolar for proper orientation

+++ tests/test_plotting.py
@@ -69,5 +69,5 @@ def test_generate_rgb_image(sample_collection):
 
     assert isinstance(color_image, np.ndarray)
     assert color_image.shape[0] == 3
-    assert color_image.dtype == np.uint8
+    assert color_image.dtype == np.uint16
     assert np.allclose(np.max(color_image), 255)
